instant movement 2

+++ CitySimulator/src/core/include/core/data_layer/data_types.h
@@ -17,7 +17,7 @@ namespace tjs::core {
 	};
 
 	struct Vehicle {
-		int uid;
+		uint64_t uid;
 		float currentSpeed;
 		float maxSpeed;
 		Coordinates coordinates;
